Split itest of KieSpringOnKarafIntegrationTest into 2 separate tests; one covering KieSpring & the other jBPM and Persistance
(cherry picked from commit 9eb38e574cb0047968fb435b8c04b2cae15b54cb)
